Move failing parser test cases into their own class
